Use enum instead of strings for Storyboards
